Calculate the gradient of line segments from a raster dataset

Calculate the gradient of line segments from a raster dataset

slope_raster(r, e = NULL, lonlat = FALSE, method = "bilinear")

Arguments

r
Routes, the gradients of which are to be calculated

e
A raster object overlapping with r and values representing elevations

lonlat
Are the coordinates in lon/lat order? TRUE by default

method
The method of estimating elevation at points, passed to the extract function for extracting values from raster datasets. Default: "bilinear".

Examples

r = lisbon_road_segments
e = dem_lisbon_raster
(s = slope_raster(r, e))

#>   [1] 3.074005e-03 1.087046e-02 4.998315e-03 1.425567e-02 5.333236e-03
#>   [6] 1.310303e-02 1.310303e-02 2.585747e-02 3.923249e-02 2.259499e-03
#>  [11] 3.420071e-03 6.338765e-02 1.475274e-02 2.300732e-02 7.431299e-03
#>  [16] 3.566200e-02 1.310303e-02 5.483172e-02 8.979176e-02 1.407353e-01
#>  [21] 1.026832e-01 2.128894e-01 1.041400e-01 1.199143e-01 1.996291e-02
#>  [26] 1.130363e-01 7.285990e-02 6.708783e-02 2.480976e-02 9.501323e-02
#>  [31] 6.387474e-02 1.325193e-02 4.288482e-02 1.310303e-02 1.310303e-02
#>  [36] 5.333236e-03 1.425567e-02 3.423468e-03 2.498694e-02 1.087046e-02
#>  [41] 3.074005e-03 1.205867e-02 4.409960e-02 7.224405e-03 4.612236e-03
#>  [46] 3.934449e-03 2.259499e-03 3.857314e-02 4.966335e-02 1.241104e-02
#>  [51] 2.455801e-02 2.054553e-02 2.521060e-02 1.347086e-01 3.611626e-03
#>  [56] 1.044272e-01 2.289246e-01 2.758316e-01 1.283347e-01 2.030014e-01
#>  [61] 1.358153e-01 7.134438e-02 1.087931e-01 1.228427e-01 1.394815e-01
#>  [66] 2.683368e-02 4.665411e-02 1.604800e-02 1.361481e-01 1.417267e-01
#>  [71] 1.433924e-01 2.123102e-01 1.802022e-01 2.265860e-02 1.071692e-02
#>  [76] 5.933725e-05 1.150703e-02 1.620129e-02 1.889174e-02 7.927565e-03
#>  [81] 1.418638e-16 3.616773e-02 2.443816e-02 3.533526e-02 1.296875e-02
#>  [86] 3.616773e-02 1.418638e-16 1.835551e-03 2.438554e-02 3.993501e-02
#>  [91] 2.765567e-02 2.317484e-02 5.063163e-02 4.483658e-02 6.371459e-02
#>  [96] 7.531101e-02 1.137188e-01 4.101872e-02 3.424646e-03 6.363231e-02
#> [101] 1.429397e-01 2.103757e-01 1.738450e-01 1.941065e-01 4.255518e-02
#> [106] 3.651447e-02 3.372936e-02 3.344048e-03 2.969840e-02 9.041050e-04
#> [111] 2.628022e-03 7.712939e-03 1.525326e-02 2.031330e-02 2.245615e-02
#> [116] 5.690689e-02 2.515642e-02 3.923249e-02 2.585747e-02 4.998315e-03
#> [121] 1.929805e-04 1.372704e-02 3.622174e-03 4.285773e-04 4.626865e-04
#> [126] 1.279684e-02 1.909044e-03 1.806464e-03 1.575317e-03 3.644256e-02
#> [131] 5.983547e-02 8.572783e-02 2.925455e-03 2.194583e-02 1.969296e-02
#> [136] 3.074005e-03 1.347086e-01 2.521060e-02 5.211203e-02 5.059354e-03
#> [141] 1.039402e-01 3.683233e-02 5.856788e-02 5.107504e-02 3.467072e-02
#> [146] 5.360840e-02 7.768261e-02 7.514460e-02 1.537696e-02 5.523761e-02
#> [151] 7.148472e-02 1.204771e-01 1.155883e-01 1.150921e-01 1.324066e-02
#> [156] 0.000000e+00 2.482997e-16 6.177143e-17 0.000000e+00 1.508745e-16
#> [161] 0.000000e+00 4.638191e-03 1.251198e-02 1.655059e-02 1.951437e-01
#> [166] 3.934819e-02 2.756013e-02 1.540712e-01 8.682767e-02 7.886121e-02
#> [171] 3.127983e-02 2.768095e-02 1.204188e-02 3.025752e-02 5.519990e-02
#> [176] 1.245545e-01 5.443024e-02 2.738194e-02 3.073584e-02 4.180824e-02
#> [181] 1.246362e-01 8.200920e-02 6.765302e-02 8.704484e-02 1.679109e-02
#> [186] 4.665411e-02 3.334219e-02 6.237032e-02 1.941099e-01 1.081655e-01
#> [191] 2.222246e-01 5.820341e-02 1.736503e-01 1.281327e-01 6.372363e-02
#> [196] 1.169481e-01 1.446041e-03 6.556451e-03 1.689554e-02 7.536771e-03
#> [201] 4.288482e-02 1.325193e-02 6.387474e-02 9.501323e-02 2.480976e-02
#> [206] 6.708783e-02 7.285990e-02 1.130363e-01 1.996291e-02 1.199143e-01
#> [211] 1.041400e-01 2.128894e-01 1.004600e-01 8.200920e-02 1.208425e-01
#> [216] 9.721340e-02 1.196593e-02 5.680142e-02 1.446867e-02 6.200052e-03
#> [221] 8.300675e-02 9.529052e-02 1.469852e-01 1.467706e-01 1.160536e-02
#> [226] 1.780237e-01 1.091320e-02 6.556451e-03 2.505247e-03 3.129335e-03
#> [231] 2.652915e-02 7.844415e-03 2.265860e-02 2.103757e-01 1.429397e-01
#> [236] 6.363231e-02 3.424646e-03 7.739929e-02 1.331082e-01 1.445820e-01
#> [241] 8.878006e-03 5.548617e-02 2.522949e-02 1.246362e-01 4.180824e-02
#> [246] 3.073584e-02 8.979176e-02 4.937218e-02 8.133036e-03 3.639028e-02
#> [251] 1.245545e-01 6.443823e-02 4.467738e-02 2.814733e-02 1.414425e-02
#> [256] 6.665666e-02 9.106105e-02 5.107504e-02 5.856788e-02 3.683233e-02
#> [261] 1.039402e-01 7.844415e-03 1.665122e-02 3.876077e-02 4.212827e-03
#> [266] 1.873147e-02 1.356266e-02 1.310303e-02 3.771554e-03 3.129335e-03
#> [271] 2.505247e-03

cor(r$Avg_Slope, s)^2

#> [1] 0.9546143