Extract slopes from xyz data frame or sf objects

Extract slopes from xyz data frame or sf objects

slope_xyz(
  route_xyz,
  fun = slope_matrix_weighted,
  lonlat = TRUE,
  directed = FALSE
)

Arguments

route_xyz
An sf object with x, y, z dimensions

fun
The slope function to calculate per route, slope_matrix_weighted by default.

lonlat
Are the coordinates in lon/lat order? TRUE by default

directed
Should the value be directed? FALSE by default. If TRUE the result will be negative when it represents a downslope (when the end point is lower than the start point).

Value

A vector of slopes equal in length to the number simple features (rows representing linestrings) in the input object.

Examples

route_xyz = lisbon_road_segment_3d
slope_xyz(route_xyz, lonlat = FALSE)

#>          1 
#> 0.09501323 

slope_xyz(route_xyz, lonlat = FALSE, directed = TRUE)

#>           1 
#> -0.09501323