Extract elevations from coordinates

Extract elevations from coordinates

elevation_extract(
  m,
  e,
  method = "bilinear",
  terra = has_terra() && methods::is(e, "SpatRaster")
)

Arguments

m
Matrix containing coordinates and elevations

e
Raster overlapping with r and values representing elevations

method
The method of estimating elevation at points, passed to the extract function for extracting values from raster datasets. Default: "bilinear".

terra
Should the terra package be used? TRUE by default if the package is installed and if e is of class SpatRast

Examples

m = sf::st_coordinates(lisbon_road_segments[1, ])
e = dem_lisbon_raster
elevation_extract(m, e)

#>  [1] 3.267991 3.266818 3.265715 3.264418 3.263222 3.262339 3.261004 3.262972
#>  [9] 3.264940 3.266908 3.268876 3.308304 3.461349 3.467100 3.439341 3.421206
#> [17] 3.403916 3.387750 3.371584 3.356889 3.342193 3.327498 3.312644 3.297117
#> [25] 3.282411 3.267716 3.253020 3.218922 3.210919 3.202916 3.194914 3.176672
#> [33] 3.157027 3.191458 3.316882 3.410792 3.469193 3.487912 3.563260 3.640330
#> [41] 3.717938

elevation_extract(m, e, method = "simple")

#>  [1] 3.370 3.370 3.382 3.394 3.406 3.419 3.104 3.116 3.128 3.140 3.152 3.233
#> [13] 3.507 3.400 3.389 3.374 3.363 3.352 3.341 3.341 3.330 3.319 3.308 3.296
#> [25] 3.285 3.274 3.263 3.252 3.252 3.241 3.230 3.219 3.208 3.315 3.451 3.451
#> [37] 3.560 3.560 3.543 3.652 3.762

# uncomment the following lines to test with terra (experimental)
# u = "https://github.com/ITSLeeds/slopes/releases/download/0.0.0/dem_lisbon.tif"
# if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
# et = terra::rast("dem_lisbon.tif")
# elevation_extract(m, et)