Calculate the gradient of line segments from a raster dataset

Calculate the gradient of line segments from a raster dataset

slope_raster(r, e = NULL)

Arguments

r
Routes, the gradients of which are to be calculated

e
A raster object overlapping with r and values representing elevations

Examples

r = lisbon_road_segments
e = dem_lisbon_raster
(s = slope_raster(r, e))

#>   [1] 2.948203e-08 1.049928e-07 4.646349e-08 1.662057e-07 5.114352e-08
#>   [6] 1.269515e-07 1.269515e-07 2.326934e-07 4.033224e-07 6.450129e-08
#>  [11] 9.491270e-08 5.910514e-07 1.335197e-07 2.098299e-07 7.309918e-08
#>  [16] 3.277216e-07 1.269515e-07 5.112778e-07 8.167900e-07 1.273213e-06
#>  [21] 1.033930e-06 2.264195e-06 1.044631e-06 1.306080e-06 2.858088e-07
#>  [26] 1.101828e-06 7.140673e-07 7.067295e-07 2.749769e-07 8.844046e-07
#>  [31] 5.886357e-07 1.238009e-07 3.867484e-07 1.269515e-07 1.269515e-07
#>  [36] 5.114352e-08 1.662057e-07 3.141131e-08 2.252912e-07 1.049928e-07
#>  [41] 2.948203e-08 1.741936e-07 4.404191e-07 1.235577e-07 1.001357e-07
#>  [46] 1.253659e-07 6.450129e-08 4.378482e-07 5.471642e-07 1.354650e-07
#>  [51] 2.648566e-07 4.695305e-07 2.715582e-07 1.744412e-06 6.287567e-08
#>  [56] 1.082368e-06 2.202623e-06 2.782562e-06 2.036142e-06 1.970080e-06
#>  [61] 2.720090e-06 7.129588e-07 1.112125e-06 1.178241e-06 1.265792e-06
#>  [66] 2.442478e-07 4.815686e-07 1.738891e-07 3.934542e-06 3.094517e-06
#>  [71] 2.468126e-06 3.906512e-06 2.017398e-06 3.132803e-07 1.012076e-07
#>  [76] 5.344741e-10 1.035569e-07 1.457931e-07 1.702694e-07 1.689494e-07
#>  [81] 1.780525e-21 3.887740e-07 2.232080e-07 3.481354e-07 1.536131e-07
#>  [86] 3.887740e-07 1.780525e-21 3.580937e-08 2.322658e-07 3.595291e-07
#>  [91] 2.540696e-07 2.904615e-07 4.560511e-07 6.655864e-07 8.108057e-07
#>  [96] 1.038519e-06 2.606659e-06 4.449564e-07 3.113458e-08 5.756507e-07
#> [101] 1.529576e-06 2.103367e-06 1.576562e-06 4.313636e-06 4.060065e-07
#> [106] 3.883512e-07 3.069144e-07 5.726603e-08 2.733030e-07 1.082223e-08
#> [111] 3.069601e-08 8.844831e-08 1.526533e-07 2.176510e-07 2.372582e-07
#> [116] 5.123210e-07 2.352346e-07 4.033224e-07 2.326934e-07 4.646349e-08
#> [121] 1.738844e-09 1.235194e-07 3.473785e-08 6.794306e-09 7.086054e-09
#> [126] 1.190326e-07 3.089683e-08 2.305145e-08 1.781576e-08 3.386903e-07
#> [131] 6.395712e-07 7.740692e-07 2.747696e-08 2.157614e-07 1.981900e-07
#> [136] 2.948203e-08 1.744412e-06 2.715582e-07 4.690934e-07 4.570934e-08
#> [141] 1.720827e-06 4.958001e-07 7.564515e-07 6.933570e-07 4.435448e-07
#> [146] 4.830895e-07 1.307358e-06 7.334462e-07 3.821527e-07 1.004095e-06
#> [151] 1.220088e-06 1.586949e-06 1.048510e-06 1.064440e-06 1.202587e-07
#> [156] 0.000000e+00 2.371438e-21 1.316730e-21 0.000000e+00 1.430375e-21
#> [161] 0.000000e+00 1.163733e-07 1.760870e-07 4.937404e-07 2.449023e-06
#> [166] 3.814840e-07 2.615932e-07 1.816155e-06 8.024127e-07 7.972038e-07
#> [171] 3.220363e-07 2.983805e-07 1.327387e-07 2.748147e-07 1.916571e-06
#> [176] 1.459918e-06 5.994643e-07 2.525053e-07 2.813632e-07 3.762023e-07
#> [181] 1.128344e-06 7.417170e-07 6.552035e-07 8.051148e-07 1.699261e-07
#> [186] 4.815686e-07 3.018545e-07 5.962307e-07 1.915773e-06 1.006557e-06
#> [191] 2.446167e-06 5.239666e-07 1.564664e-06 1.252903e-06 6.907348e-07
#> [196] 1.376381e-06 1.657874e-08 5.951644e-08 2.052076e-07 1.296014e-07
#> [201] 3.867484e-07 1.238009e-07 5.886357e-07 8.844046e-07 2.749769e-07
#> [206] 7.067295e-07 7.140673e-07 1.101828e-06 2.858088e-07 1.306080e-06
#> [211] 1.044631e-06 2.264195e-06 1.040684e-06 7.417170e-07 1.091681e-06
#> [216] 8.913548e-07 1.357584e-07 5.187953e-07 2.865644e-07 5.797358e-08
#> [221] 7.592161e-07 8.620683e-07 1.323121e-06 1.414720e-06 1.105911e-07
#> [226] 1.683123e-06 1.026096e-07 5.951644e-08 2.859523e-08 5.201032e-08
#> [231] 2.567781e-07 9.510530e-08 3.132803e-07 2.103367e-06 1.529576e-06
#> [236] 5.756507e-07 3.113458e-08 1.669466e-06 1.534675e-06 1.365440e-06
#> [241] 8.138565e-08 5.070468e-07 2.287617e-07 1.128344e-06 3.762023e-07
#> [246] 2.813632e-07 8.167900e-07 4.462902e-07 7.321778e-08 3.996923e-07
#> [251] 1.459918e-06 2.355208e-06 4.029401e-07 2.535103e-07 1.403490e-07
#> [256] 6.828694e-07 9.539718e-07 6.933570e-07 7.564515e-07 4.958001e-07
#> [261] 1.720827e-06 9.510530e-08 2.087078e-07 3.594839e-07 9.020427e-08
#> [266] 1.728241e-07 1.618832e-07 1.269515e-07 6.248451e-08 5.201032e-08
#> [271] 2.859523e-08

cor(r$Avg_Slope, s)^2

#> [1] 0.7717239