Calculate the gradient of line segments from a raster dataset

Calculate the gradient of line segments from a raster dataset

slope_raster(
  r,
  e = NULL,
  lonlat = FALSE,
  method = "bilinear",
  fun = slope_matrix_weighted,
  terra = has_terra()
)

Arguments

r
Routes, the gradients of which are to be calculated

e
A raster object overlapping with r and values representing elevations

lonlat
Are the coordinates in lon/lat order? TRUE by default

method
The method of estimating elevation at points, passed to the extract function for extracting values from raster datasets. Default: "bilinear".

fun
The slope function to calculate per route, slope_matrix_weighted by default.

terra
Should the terra package be used? TRUE by default if the package is installed and if e is of class SpatRast

Examples

# r = lisbon_road_segments[239, ]
r = lisbon_road_segments
e = dem_lisbon_raster
(s = slope_raster(r, e))[1:5]

#> [1] TRUE

#> [1] 0.003074005 0.010870459 0.004998315 0.014255667 0.005333236

cor(r$Avg_Slope, s)

#> [1] 0.9770436